+++ inc/storefront-functions.php
@@ -82,10 +82,10 @@ if ( ! function_exists( 'storefront_header_cart' ) ) {
 					</ul>	
 				</div> 
 				<?php if ( !is_checkout() && !is_cart() ) { ?>
-				<div class="drawer-side z-10 h-full">
+				<div class="drawer-side z-10 h-full overflow-x-hidden">
 					<label for="my-drawer-4" aria-label="close sidebar" class="drawer-overlay"></label>
-					<ul class="menu p-4 w-80 min-h-full bg-neutral-50 text-base-content">
-					<?php the_widget( 'WC_Widget_Cart', 'title=' ); ?>
+					<ul class="menu p-4 w-80 min-h-full bg-white text-base-content">
+						<?php the_widget( 'WC_Widget_Cart', 'title=' ); ?>
 					</ul>
 				</div>
 				<?php } ?>
@@ -125,3 +125,15 @@ function receitasdoparaiso_customize_register( $wp_customize ) {
 	// $wp_customize->remove_panel( 'widgets');
 }
 add_action( 'customize_register', 'receitasdoparaiso_customize_register', 20 );
+
+
+/**
+* remove WooCommerce tabs from product page
+*/
+
+function my_remove_description_tab( $tabs ) {
+  unset( $tabs['description'] );
+  return $tabs;
+}
+
+add_filter( 'woocommerce_product_tabs', 'my_remove_description_tab', 11 );