+++ inc/woocommerce-functions.php
@@ -89,22 +89,6 @@ function cxc_display_single_sale_price_after_sale_price() {
 
 add_action( 'woocommerce_single_product_summary', 'cxc_display_single_sale_price_after_sale_price', 15 );
 
- /**
-*  WooCommerce Check if User Has Purchased Product
-*/  
-function bbloomer_user_logged_in_product_already_bought() {
-    global $product;
-    if ( ! is_user_logged_in() ) return;
-    if ( wc_customer_bought_product( '', get_current_user_id(), $product->get_id() ) ) {
-        echo '<div role="alert" class="alert">
-            <svg xmlns="http://www.w3.org/2000/svg" class="stroke-current shrink-0 h-6 w-6" fill="none" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" /></svg>
-            <span>Este produto já foi adquirido</span>
-        </div>';
-    }
-}
-
-add_action( 'woocommerce_after_shop_loop_item', 'bbloomer_user_logged_in_product_already_bought', 30 );
-
 function bbloomer_hide_add_cart_if_already_purchased( $is_purchasable, $product ) {
     if ( wc_customer_bought_product( '', get_current_user_id(), $product->get_id() ) ) {
        $is_purchasable = false;
@@ -114,5 +98,170 @@ function bbloomer_hide_add_cart_if_already_purchased( $is_purchasable, $product 
  
 add_filter( 'woocommerce_is_purchasable', 'bbloomer_hide_add_cart_if_already_purchased', 9999, 2 );
 
+ /**
+*  WooCommerce new product badge 90 days
+*/   
+function receitas_do_paraiso_new_badge_shop_page() {
+   global $product;
+   $newness_days = 30;
+   $created = strtotime( $product->get_date_created() );
+   if ( ( time() - ( 60 * 60 * 24 * $newness_days ) ) < $created ) {
+      echo '<span class="itsnew onsale bg-secondary text-white">' . esc_html__( 'Novo!', 'woocommerce' ) . '</span>';
+   }
+}
+
+add_action( 'woocommerce_after_shop_loop_item_title', 'receitas_do_paraiso_new_badge_shop_page', 3 );
+
+ /**
+*  WooCommerce Check if User Has Purchased Product
+*/  
+// function bbloomer_user_logged_in_product_already_bought() {
+//     global $product;
+//     if ( ! is_user_logged_in() ) return;
+//     if ( wc_customer_bought_product( '', get_current_user_id(), $product->get_id() ) ) {
+//         echo '<div role="alert" class="alert">
+//             <svg xmlns="http://www.w3.org/2000/svg" class="stroke-current shrink-0 h-6 w-6" fill="none" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" /></svg>
+//             <span>Este produto já foi adquirido</span>
+//         </div>';
+//     }
+// }
+
+// add_action( 'woocommerce_after_shop_loop_item', 'bbloomer_user_logged_in_product_already_bought', 30 );
 
 
+/**
+ * Hide loop read more buttons for out of stock items 
+ */
+if (!function_exists('woocommerce_template_loop_add_to_cart')) {
+	function woocommerce_template_loop_add_to_cart() {
+		global $product;
+		if ( ! $product->is_in_stock() || ! $product->is_purchasable() ) return;
+		wc_get_template('loop/add-to-cart.php');
+	}
+}
+
+
+// Add custom button for already purchased products on single product page
+function add_custom_button_for_purchased_product($price_html, $product) {
+    if ( wc_customer_bought_product( '', get_current_user_id(), $product->get_id() ) ) {
+        if ( is_product() && ! is_archive() ) {
+            $price_html .= '<div role="alert" class="alert my-4 bg-neutral-50 font-medium w-fit h-12 leading-none flex">
+            <svg xmlns="http://www.w3.org/2000/svg" class="stroke-current shrink-0 h-6 w-6 text-success" fill="none" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" /></svg>
+            <span>Este produto já foi adquirido, descarregue <a class="text-underline text-primary" href="/minha-conta/downloads/">aqui</a>.</span>
+        </div>';
+        } else {
+            $price_html .= '<div role="alert" class="alert my-4 bg-neutral-50 font-medium w-fit h-12 leading-none flex mx-auto">
+            <svg xmlns="http://www.w3.org/2000/svg" class="stroke-current shrink-0 h-6 w-6 text-success" fill="none" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" /></svg>
+            <span>Este produto já foi adquirido</span>
+        </div>';
+        }
+    }
+    return $price_html;
+}
+
+add_filter('woocommerce_get_price_html', 'add_custom_button_for_purchased_product', 10, 2);
+
+
+/**
+
+ * WebAppick - Set session variable on page load if the query string has coupon_code variable.
+
+ */
+
+function ts_get_custom_coupon_code_to_session() {
+
+if( isset( $_GET[ 'coupon_code' ] ) ) {
+
+     // Ensure that customer session is started
+
+     if( !WC()->session->has_session() )
+
+         WC()->session->set_customer_session_cookie(true);
+
+     // Check and register coupon code in a custom session variable
+
+     $coupon_code = WC()->session->get( 'coupon_code' );
+
+     if( empty( $coupon_code ) && isset( $_GET[ 'coupon_code' ] ) ) {
+
+         $coupon_code = esc_attr( $_GET[ 'coupon_code' ] );
+
+         WC()->session->set( 'coupon_code', $coupon_code ); // Set the coupon code in session
+
+     }
+
+}
+
+}
+
+add_action( 'init', 'ts_get_custom_coupon_code_to_session' );
+
+/**
+ * Apply Coupon code to the cart if the session has coupon_code variable.
+ */
+
+function ts_apply_discount_to_cart() {
+
+    $coupon_code = WC()->session->get( 'coupon_code' );
+
+    if ( ! empty( $coupon_code ) && ! WC()->cart->has_discount( $coupon_code ) ) {
+        WC()->cart->add_discount( $coupon_code ); // apply the coupon discount
+        WC()->session->__unset( 'coupon_code' ); // remove coupon code from session
+    }
+}
+
+add_action( 'woocommerce_before_cart_table', 'ts_apply_discount_to_cart', 10, 0 );
+
+
+function woocommerce_coupons_enabled_checkout( $coupons_enabled ) {
+    global $woocommerce;
+    if ( ! empty( $woocommerce->cart->applied_coupons ) ) {
+        return false;
+    }
+    return $coupons_enabled;
+}
+
+add_filter( 'woocommerce_coupons_enabled', 'woocommerce_coupons_enabled_checkout' );
+
+
+function woocommerce_store_notice_alert($text) {
+    global $woocommerce;
+
+    $applied_coupons = WC()->cart->get_applied_coupons();
+    $discount_status = do_shortcode('[rp_options key="discount_status"]');
+    $discount_percentage = do_shortcode('[rp_options key="discount_value"]');
+ 
+    if( count( $applied_coupons ) > 0 ) { 
+        $text = str_replace(
+        array('<p class="woocommerce-store-notice demo_store">', '</p>', 'Dismiss'), 
+        array('
+        <div class="hello"><p class="woocommerce-store-notice demo_store">', 
+        "Cupão de desconto de $discount_percentage% aplicado com sucesso.</p></div>", '(close svg icon here)')
+        , $text);
+    } else {
+        $text = str_replace(
+            array('<p class="woocommerce-store-notice demo_store">', '</p>', 'Dismiss'), 
+        array('
+        <div class=""><p class="woocommerce-store-notice demo_store">', 
+        "<span class='font-semibold mr-6'>Clica para obteres cupão um desconto de $discount_percentage% em todos os Ebooks, válido em apenas uma compra.</span><a href='/carrinho/?coupon_code=bundle10'><button class='btn btn-primary'>Aplicar $discount_percentage%</button></a> </p></div>", '(close svg icon here)')
+        , $text);
+    }
+
+    if($discount_status) {
+        return $text;
+    }
+    
+    return null;
+}
+
+add_filter('woocommerce_demo_store', 'woocommerce_store_notice_alert', 10, 1);
+
+
+function display_store_notice_on_shop_page_only() {
+    if ( is_shop() || is_cart() ) {
+        add_action( 'wp_footer', 'woocommerce_demo_store', 10 );
+    } else {
+        remove_action( 'wp_footer', 'woocommerce_demo_store', 10 );
+    }
+}
+add_action( 'wp', 'display_store_notice_on_shop_page_only' );