+++ template-parts/home/stats.php
@@ -3,15 +3,15 @@
     <dl class="grid grid-cols-1 gap-x-8 gap-y-16 text-center lg:grid-cols-3">
       <div class="mx-auto flex max-w-xs flex-col gap-y-4">
         <dt class="text-base leading-7 text-white">Subscritores</dt>
-        <dd class="order-first text-3xl font-semibold tracking-tight text-white sm:text-5xl">+250 mil</dd>
+        <dd class="order-first text-3xl font-semibold tracking-tight text-white sm:text-5xl">+<?php echo do_shortcode('[rp_options key="rp_subscribers"]'); ?></dd>
       </div>
       <div class="mx-auto flex max-w-xs flex-col gap-y-4">
         <dt class="text-base leading-7 text-white">Visualizações</dt>
-        <dd class="order-first text-3xl font-semibold tracking-tight text-white sm:text-5xl">+30,000,000</dd>
+        <dd class="order-first text-3xl font-semibold tracking-tight text-white sm:text-5xl">+<?php echo do_shortcode('[rp_options key="rp_views"]'); ?></dd>
       </div>
       <div class="mx-auto flex max-w-xs flex-col gap-y-4">
-        <dt class="text-base leading-7 text-white">Vídeos</dt>
-        <dd class="order-first text-3xl font-semibold tracking-tight text-white sm:text-5xl">+400</dd>
+        <dt class="text-base leading-7 text-white">Receitas</dt>
+        <dd class="order-first text-3xl font-semibold tracking-tight text-white sm:text-5xl"><?php echo $count_recipes = wp_count_posts( 'receitas' )->publish; ?></dd>
       </div>
     </dl>
   </div>